[mangasee] Workaround anti-scraper protection
Mangasee changed a var name to (weakly) deter scrapers because load
became too high. As a result, the regex had to be updated to get the
domain of the image. The change was supposed to be temporary but just
added a resilient solution to the old and current var name.