Moved methods intended to be overridden from Server to GenericServer
The server class was broken up to better indicate what needs to be
overridden in order to streamline the process of adding new servers.

And to that end unused and implied variables were removed/cleaned up.

has_login is autodetected based on if the login func was overridden
Unused server properties were removed including has_gaps which was
used for some obscure tests.