Better handle tracker and server differing in punctuation
Trackers and a given server often have different punctuation for series
which causes problems when trying to use the load subcommand. A prior
attempt to solve this was to search for the fist word which led to really
generic searches (like for "the") which did more harm than good. Now we'll
just look for the first part with punctuation which is expected to perform
better in most cases.