[jnoveclub] Actually enforce deleting part data
One of the requirements to use the J Novel Club's api is to remove part
date when it is no longer available. The plan had been to reply on the
synced_removed setting, but it is actually hard to tell if a part is
available for not without actually downloading it. So let's just delete
any saved parts after some fixed time.