Standardized m3u8 downloading
ANIME servers have the option of proving a m3u8 url or a mp4. Either is
fine for streaming, for downloading m3u8 require more work as the actual
ts files need to be extracted. However this is basically the same
process for all the servers so the code can be shared.