local_server: Use standard file hierarchy
Removed support of the collapsed file hierarchy for the local server.
Now all chapter related info must reside in a directory rather than in
the bare media directory. This also prevents the metadata from being
mistake and chapter data.

When importing a directory, the directory just servers as the name which
the files within it will be imported under. If the files of the
directory were specified themselves, the name would be determined from
the individual file names instead.