update: Remove chapter data that is no longer available
Changed default behavior of update to remove chapter data if it is no
longer available on remote server and it is not already downloaded. The
rationale being that the data may have changed ids and having both would
cause us to fail to consume the intended data.

This was the default for certain servers, but that flag no longer
exists. If the old behavior is desired, a new setting keep_unavailable
has been added.