Allow overriding server specific settings from env
Media specific settings can now be override from env variables if they
are specific to a media type. This allows changing the default view for
media types without modifying the settings file.

To take advantage of this either the general var is None or its value
has been overridden by the environment. When either of these conditions
are met, "_" + MediaType::name will be appended to the env var name for
each media type and the proper server specific field will be set if the
env var exists.