Don't crash on empty chapter_metadata
If a server stops providing media info after it has already provided it,
an empty chapter_metadata.json will exists. However, an empty file is
not valid json, so attempting to parse it will cause an error to be
thrown. We now treat this error the same as the file not existing and
ignore it.