Add support for simple media servers
A bare-bones media servers can now be setup from the config file. All
the server has to do is real its directory structure (typical behavior
if index pages aren't set) and with some minor config, can be used like
any remote server. This allows for storing media on one machine and
accessing it from another and it is compatible with the format for
imported media. So one could
1. Import some anime on machine A
2. Start the webserver and point it to XDG_CONFIG_HOME/amt
3. Specify the path to local_anime
4. Access the media like one could if it was publicly hosted

This commit also standardized import names for media such that importing
via `import` and manually dropping the file in the correct position will
yield the same names/ids.