remote: Fix problems relation to authentication
There was a problem with test coverage and RemoteServers have been
disabled for a while. This has been fixed along with some bugs that were
hidden by the lack of test coverage including not being able to access
any media when authentication was used.

Now local and remote server need to have 100% code coverage when doing a
quick test and tests have been updated to enforce this.