vrv: Allow vrv to process crunchyroll beta links
Turns out the crunchyroll beta uses the same media ids as VRV. The
original plan was to implement a new server for the beta website so its
urls could be streamed. But since the ids are the same and the url is in
the same format as VRV, VRV just gained the ability to stream the urls.