settings: change how passwords are stored
Password save/load commands used to accept the server_id by position.
Now they do so by name. Also username is not longer passed it as input
but instead subsisted as part of the format string. This also removed
the need for an explicit credentials_separator which has now been
replaced with a regex to determine how to slit the input.

This also resolves a problem where "\n" couldn't be used as a
credentials_separator because of how settings are read.

Also handle the case where the external programs give multiple
username/passwords by only accepting the first ones.