mangasee: Share cookies with cloudscraper sessions
Due to a bug in upstream cloudscraper cookies wouldn't be shared with
the normal session if at initiation the cookie jar was empty. We don't
load load cookies until after servers have been created so it would
always be empty. The main affect of this was login cookies not be saved
for humblebundle.

Until upstream is fixed, we manually set the cookie jar for
cloudscraper. However upon doing so it was observed that we ended up
clobbering headers, so the workaround of only setting cookies may
actually be the correct thing to do.

We also ensure there is only 1 cloudscraper server