Stopped relying on isort's arbitrary and opinionated import sorting.
`imports` are before `from` and system imports are before local. There's
a space between each group and Within these groups everything is sorted.

The real reason for this change is because isort has too many
dependencies to build and in general is too complicated. Can achieve
basically the same thing in about a dozen lines of vimscript.

Also stopped validating import sorting on push