download_unread_chapters: download in sane order
Individual pages are normally downloaded in parallel. However,
download_unread_chapters tried to download chapters in parallel too
which often overwhelmed the server. It has now been changed to only
download chapters in parallel iff if is for a different media_data.

This also ensures that the lowest number chapter will be the first one
downloaded for that media (unless shuffle is used).