download: Added option to ignore errors when downloading chapters
It was observed that Vizmanga will sometimes report an extra page and
said page can't be downloaded. This new force flag added to the consume
and download family subcommands will now allow such errors to be
ignored and not block ones progress of an entire series.

It should however be used with caution because all errors are ignored
after the page metadata has been loaded (so doesn't count internet
errors/premium errors). After that point the chapter will always be
marked downloaded and once mark it won't be tried again.