mangadex: update to subtle api changes
Removed a chapter for the key "data" in chapter attributes. Previously
chapters without this check were skipped. This field no longer seems to
be present and it is unclear why it was needed so it has just been
removed.

It was also required to change how chapters were downloaded. Now there
is one less url to fetch.

Also allow stream_index to toggle between normal and dataSaver images.