crunchyroll: don't hardcode media list
Due to Crunchyroll's bot protection and apparent lack of public api to
search/list media, the media list was hardcoded. Now it we rely on
cloudscraper to bypass the bot protection and scrape the webpage for the
media list. There doesn't seem to be a perfect page to scrape so
currently we scrape and combine the alphabetical list and the popular
page which generates a list virtually equivalent to what was hardcoded.
Some media appear "lost" and it is believe they are now longer available
on the platform.