hidive: added new server
Finally support the last of the DRM-free major US anime streaming
providers. There were several challenges with HiDive:

We need to have the write cookies and headers to even watch the stream
and not to just get the stream info. This meant passing the cookie file
and specifying HTTP headers to the actual player.

Also needed to handle the fact that the subtitles were 5s off for
whatever reason in addition to having the same problem as vrv
(unsurprising since they had the same source) with multi-line subtitles
appearing out of order.