crunchyroll: explicitly require cloudfare
The Crunchyroll manga server only needs cloud scraper to search/list media
but not to view them. Making cloud scraper a hard requirement seemed
heavy-handed, but the previously solution of just relying on a different
server to init the cloud scraper session was... hacky and didn't work
for tests. Now crunchyroll explicitly requires it.