dbmultiverse: update in response to upstream changes
So dmbultiverse changed the name of the one of  the classes  that held all the
chapters. This caused is not being able to find any chapters. They also
added an additional class and that caused us to only be able to find
some servers which is really hard to catch in tests.