server: Enabled per server threads
Some servers like funimation have to make multiple queries when
listing/searching for media. A while back these queries were parallelized
and now the exact number of threads can be controlled on a per server
level.

Severs that use multiple threads now set the multi_threaded which allows
multithreaded tests to be aware of them (and not spam remote servers)