Disable servers if they fail to be instantiated.
This fixes a bug where if a cloudscraper enabled server isn't disabled,
it will prevent MediaReader from being instantiated because it would
fail to import cloudscraper. Normally import issues are resolved when
the class is first imported, but this specific one is an exception since
only the constructor needs to explicitly access it. (The other
exceptions are for optional features.)