Make always_use_cloudscraper and fallback_to_insecure_connection play nice
cloudscraper apparently doesn't like it when verification is explicitly
disabled and won't attempt the connection. It would error with:

ValueError: Cannot set verify_mode to CERT_NONE when check_hostname is
enabled.

Honestly, servers that need validation disabled don't use cloudfare but
this can be a problem if one just sets always_use_cloudscraper global
and needs fallback_to_insecure_connection for a specific server. The
immediate workaround is to just retry in a normal, non-cloudscraper
session. Cookies should be shared between the 2 so it should be fine.

In addition, a warning message is prevented to clearly indicate the
issue. Tests have also been added to validate the use of
always_use_cloudscraper with all other servers.