search: Potentially batch search terms
When searching for media manually or when trying to find a match from a
tracker, there were problems if the name wasn't an exact match. To
workaround around this, the given search term was broken up and the
search was preformed on all parts upon failures. This was very slow.

Now each server has control of the search term and go to the breakup
themselves. This is significantly faster for servers who just cache a
media list and no worse for those that do manual fetching.