wlnupdates: removed server
wlnupdates had an awkward api and since it was just a aggregator of
other sources wasn't too useful. Each source had to be implemented
independently so it may be more practical to just make each of them
their own server. Don't want dozens of generic webcomic servers so will
probably revive this server with a whitelist of "support" sites. There's
nothing I'm really dying to read that would need this so it'll be a
low priority.