Use external links to link media to tracker
Trackers like Anilist have links to official sources. We can use these
links to guess the association between server and tracker. This isn't
100% as the links are generally to the media and may not be specific to
the season so it isn't a 1-1 match. We will still prompt users to select
a match and will fallback to the existing methods if we can't find them.
Checking already-added media still has the highest priority as that's
cheapest to do.