Allow tracker to report the time the next episode is available
Trackers like Anilist know the time the next episode is scheduled to
air, so we can use that info and report it in the list subcommand. Note
that this may be slightly off from when the episode is available to the
user since some servers may incur/induce delays for varying reasons.

This will only be useful for Anime since they generally have well know
release schedule which works out fine because Anime Servers are the
hardest to get that info out of.