test: Added tests to cover the special subtitle handling
With the introduction of the hidive server, the special subtitle handing
used for vrv was moved to server.py so it could be shared. This meant
that we weren't getting proper coverage in server.py (we have relax
coverage requirements for servers requiring logins for automated tests).

Modified TestAnimeServer to report "download" subtitles that were
extracted from real data from HiDive. This covers the offset and the
subtitle inversion. Note that we can't validate correctness since it
just affects appearance/subtitle sync but simple coverage should be
sufficient to catch stupid bugs.