args: propagate function return code to the process's exit code
When a function fails without erroring, we'll now exit with a non-zero
error code instead of just silently ignoring the error. The behavior for
errors remains unchanged.

The criteria for a failures is a little vague. One type that will always
cause a non-zero exit code is the inability to start the operation, like
giving the stream command a bad url.