search: Use string distance when sorting search entries
Instead of guessing closeness by by diffing the resulting media name
with the search term, use an actual string distance function. We may end
up searching based on multiple terms to we sum the score for the best
match and the literal term.

Also we now return the score in server::search and use that to sort
media across servers. In addition limit is now applies to the output so
the screen isn't spammed with entries. As a result all searching should
yield more accurate results in a more friendly manner.