crunchyroll: get new session id on failed login
There was a bug where a new session id wouldn't be generated when trying
to login in. This mean that if your session had expired, login would
just fail. To workaround this, one could just call `update` which would
result in new session id being generated; after that subsequent
attempts to login would work.

But now the login method has been fixed to detect an expired session and
generate a new one.