Attempt to auto-login when viewing mature content.
Most of the anime servers would refuse to play "mature" content unless
the user logged in. To workaround all the user used to have to do was
run `amt login`.

Now they can do nothing and amt will login on its own just like it does
when playing premium content.

The reason we didn't didn't this before, is because it isn't trivial to
generally detect mature shows for all servers. Even now instead of
detecting it, we just retry if we failed to play the media.