Employ exponential backoff on retryable failures
If an error is encountered and it is one of settings.status_to_retry,
we'll sleep before retrying the request. And for every subsequent
failure we'll sleep exponentially longer according to the new backoff
setting.

Note this is similar to the retry mechanism built into requests which
was removed to allow for special handling when failing to setup the
connection.