Auto print output of command
Now functions won't auto print their results. Instead they will be
passed back to the caller. This would make it easier to integrate with a
GUI wrapper (which probably wants to write the output to the screen) and
makes testing easier. When args.py invokes one of these functions, it
will auto print the results.