stream: don't assume globally unique chapter ids
There was a mild optimization that made it moderately faster to stream
shows that were already registered. We got the chapter id and checked to
see if any registered media had such a chapter. If there was we could
save the subsequent check to load the media.

However, it isn't guaranteed that chapter ids would be unique outside its
media. In fact hidive, has very non-unique chapter ids like s01e001. In
such cases if you tried to treat season 1, episode 1 of any hidive show
when you already season 1 of some other hidive show registered, you'd
load that show instead of the desired one.

Instead of amending the api to workaround this, the optimization has
just been removed. It should be rare that one would stream a show they
already had registered as it would be faster to just use the number
directly and if they wanted to, they wouldn't lose that much time by
re-doing the lookup.

A side-effect of this change is that streaming will no longer affect the
watch count if the show was already registered. This could easily be
avoided, but this also seems like an improvement. In case the old
behavior is desired, a -r/--record option has been added to the stream
command.