stream: Handle streaming tracked but recently updated media
Previously, if one wanted to stream the latest chapter of registered
media without explicitly updating and fetching the latest chapter data,
the media fail to be found. Ironically, if the media wasn't registered,
it could be streamed normally. Now we'll update the media if we can't
find the requested chapter to avoid this case