crunchryoll_anime: Avoid race when trying to get a new Crunchyroll session
Finally fixed the oldest known bug in amt. Many crunchyroll requests
require a session id. This id is returned in a cookie buy may become
invalid before the cookie expires. So we don't know if it is valid so we
just attempt to use it and if we get an error, get a new id.

Now when trying to update several crunchryoll anime, we would end up
sending multiple requests which would all fail and then again try to get
a new session id multiple times which interfered with each other. Now we
serialize on the first request before letting all the rest go through