jnovelclub_parts: Don't autodelete latest streaming part
If a j_novel_club part is currently streaming and hasn't completed don't
auto delete it if downloaded. We auto delete downloaded parts after a
week to comply with the ToS. However, because it is hard to tell if a
part is currently available for download or not. But since the latest
volume is always available at least until it finishes streaming we can
safely not delete it.