Don't auto print dict in cli
The change to print the return values from subcommands was to save each
subcommand from printing the value themselves. We only did this for
numeric and list values when in made sense. We never did it for dicts
because that generally isn't useful, however that ability was added as a
side effect and is just distracting. Now we skip print dict values if a
subcommand return them subcommands