crunchyroll: Update to new crunchyroll
Crunchyroll beta is no longer in beta and this broke how we lookup the
session id. Luckily we can go to the manga homepage and get our session
id and everything else seems to continue to work (for now).

Still the former beta site is basically a reskin of vrv so that still
works and has the same content.