Removed MatureContentException
Sometimes playing a non-premium media would fail because it was "mature
content" and the user wasn't logged in. It wasn't worth it to detect if
a media fit into this category so we assumed that if getting the media
failed it was for this reason. However another valid reason would be the
user isn't premium and our premium detection failed.

So now we just catch KeyError, login if needed and if we still fail just
propagate the error with a descriptive log line