get_media_list: remove limit keyword from params if it wouldn't save http requests
It was becoming difficult to add optional params to get_media_list
because every file would needed to be updated. Now servers just
specify the args the use and use **kwrags for the rest. Additionally had
servers drop the limit param if it wouldn't save on requests. Reducing
the result list can be done after sorting to possibly yield more
accurate results.