torrent: Make torrents first class citizens
Torrents used to require a lot of special handling due to how different
they are from require streams. An abstraction layer has been so they can
act like regular http servers.

The abstraction layer is provided by user specified commands and if they
are set the servers are disabled. These commands can additionally be
used to stream a torrent.

The net result is that TorrentHelpers and the autoimport command have
been removed. A new torrent server has been added and the nyaa server
has been modified to extend from it