Add --batch-size option
This carries on the spirit of the bundle command which was removed
earlier. It existed to provide a way to view multiple chapters at once
which was reminiscent of reading a magazine. However, bundle required
too much special handling so this should be a much lighter weight
alternative.

Also modified get_children to not use wildcards but return a list of the
pages.