crunchryroll: search improvements
Stop overriding search_helper and override search_for_media like
everyone else. This lets Crunchryroll take advantage of search caching.
The reason we didn't was to reduce the risk of duplicate work since the
same search terms may result in the same media and we didn't
have to look them up multiple times. However, we could just cache each
individual request to avoid the penalty.

Another side effect of overriding search_helper is that the server
treat each term with the same weight when in reality terms earlier in
the list should be given preference. This in combination with the limit
param led crunchy to return bad results even when it original found
the desired media.