torrents: refactor torrent spawning
Instead of spawning both the torrent cmd and the regular viewer and
setting up the pipe, just spawn torrent cmd and tell it the viewer
command to run. This can enable torrent command know when the viewer
dies without having to write to it. So now it can be a good citizen and
continue seeding until the video finishes playing and also close its
stdout when it has outputted everything so the viewer actually knows it
is done.