args: Only allow batch size with viewer command
The addition of the batch size option was meant of replace the defunct
bundle command. Both were only intended for manga, but --batch-size
could be used for the play and consume args. Weather this actually
worked or not really depended on the custom program used, the player
command and the media itself.

The main issue stems from where the command should run since certain
commands make make assumptions on where to load data (like subtitle for
anime or images for novels). Additionally torrents may store a cache
file relative to the current directory.

Additionally it makes even less sense to try to consume media of
different types with --batch-size. Instead of preventing this, the help
just says not to do it.