crunchyroll_anime: cache updates for seasons that have finished airing
Subsequent requests to get season data for Crunchyroll anime seem to get
exponentially slower. And in practice we often won't get new data. We
only expect new data if the season is currently airing so in all other
cases just cache the requests.

Note that to detect if the season is currently airing, we just look at
the gab between the last episode data and now and if it is over some
configurable threshold we assume the season has finished. This isn't
perfect but probably good enough for most cases. And to cover the case
where there was some gab in the airing schedule, we'll refresh the cache
entry every week. And if the user wants a faster response, they can
clear the cache themselves with existing options.