hidive: properly handle Hidive logo
Previously Hidive videos were composed of the logo and the actual video.
However, mpv misbehaved after playing the logo and skipped 5s (the
duration of the logo) into the actual video. So we lost that content but
this also led our subtitles to be 5s off.

We hacked around this by manually offsetting our subs by 5s. And this
kinda worked, however this problem only occurred when streaming a
video. So if the video was downloaded, then offsetting messed up the
subs.

So now instead of that hack, we just split the video into 2 parts and
give each to the player. With mpv at least, this allows for proper
behavior.