remote_server: rewrite
Changed how remote_server determines what is a media/chapter.
Now special json files can server as markers to denote a media directory
and/or chapter directory. If these files have non-empty json, then they
will be taken as an index of subdirectories which saves us the trouble
of walking the entire tree.

The names of the special metadata files are the same as the what any
normal server uses. So one could point their remote webserver to a
valid amt state directory and all downloaded media would be exposed