server: Relogin if needed to list/search/update media
Some servers only report medium users have purchased so they won't show
anything if the user isn't logged in. Previously one would have to
manually log in or so some action (like potentially playing a video) that would
cause the user to relogin. If they didn't an attempt to list/search
would yield no results or an error depending on the server. Now we
relogin if needed before attempting to list/search/update