crunchyroll_anime: Actually use Crunchyroll to stream Crunchyroll urls
Since Crunchyroll and vrv share media ids, we used the vrv server to stream
urls. This was done to avoid rewriting the Crunchy anime server code
after the beta site went GA. The old api still functioned so this was
passable. However, streaming urls had some problems.

Specifically for this usecase, the convert flag was added to the stream
command. It allowed us to pass in a vrv url (or a crunchyroll url),
match the vrv server and then search for a similar media across all
servers. However this could end up being really slow and inefficient
depending on how the searches go.

So now we probably extract the media/chapter id from the url in the
crunchyroll_anime server like we should have done in the first place. A
side effect of this change is that the chapter_id changed which prompted
us to bump the server version.