stream: Make streaming urls more robust
If a url is given for streaming but is unrecognized, follow the
redirects and see if the destination is streamable.

Add more tests to verify that the protocol, www prefix, trailing slash
etc aren't required to match a url.

crunchyroll & mangadex streaming_regexes have been updated to be more
flexible

Test streaming urls have been resorted and updated