humblebundle: Remove get_prompt_for_input
This was a terrible idea. Humblebundle required 2 factor identification
to login, so code was added to give the user a chance to enter the code
that would be emailed to them. However, they since required a captcha
before the code was sent which made this "feature" useless.

In addition, we'll attempt to login atomically so we could end up
prompting a user for a code as a result of a general search which may
appear as amt hanging.

So we are just scrapping this idea. Users who want to use this server
should just manually log in and copy cookies.