args: Add import-cookies command
Previously we there was a list of read-only cookies files there would be
added to the session. The idea was to be able to share cookies with a
browser that saved cookies to same known location. However since the
file was readonly we'd keep replacing our cookies with the external one
which could get us into a situation where we couldn't get out of
without manual intervention. For example if login session expired, we'd keep
overriding our potentially valid cookies with the invalid ones every
boot.

There was also a problem in that we'd import cookies from a bunch of
irrelevant domains and drastically increase the size of our cookie file.

Now we added the import-cookies command that will import cookies related
to the specified server from a given file.