settings: add cookies field
There a few servers with broken/incomplete login code generally due to
needing to complete a captcha in order to login. The current way to use
these servers it to login via some external means (like a browser) and
copy your cookies over.

This commit adds another method. Instead of blindly copying all cookies
in a domain, one can now copy cookies from a json file. The idea is to
be able easily to share login in state with our cli programs. We import
the state as cookies as a convenient way to let the needed server access
the information.