Preserve chapter_info after media has been re-enabled
When re-enabling a disabled media, we don't immediately populate the
chapter info. This will lead to the chapters field being empty thus
causing us to persist that there is no chapter info and overwrite
anything we had saved.

Now we explicitly load chapter info when we are re-enabling a media.