Don't update progress for untracked media
Previously, if we tried to sync untracked media, the progress would be
updated. This meant that if we later started tracking, a subsequent sync
would be a no-op since we think we are already update to date with the
track. Now we won't update progress if the media isn't tracked.