crunchyroll_anime: stop using old crunchyroll ids
Crunchyroll finally removed the ability to get media info from old
non-guid ids. So we finally updated to used their guids.

Note that this commit changes the media ids from all media on the
crunchyroll_anime server. We'll automatically convert the media to the
new format which will make the state incompatible with any older
versions