Support downloading subtitles in all possible languages
Previously we'd only download the first subtitle file for a given type.
Now we can download all of them given they are have a unique lang, ext
pair. It is up to the player to parse the language from the file name.

Note mpv pre version 0.35.1 won't be able to parse the language when
streaming a url.