mangasee: don't report chapters and volumes
Previously if a media entries for both individual chapters and compiled
volumes, it would list both. This led to a jarring UI as there would be
"duplicates" and there could appear to be significantly more volumes
than reality.

Now we detect this behavior and will only show volumes if both are
present.