server: remove need_cloud_scraper
Most servers don't need cloudscraper for everything. The ones that
"depend" on it need it for searching but can generally stream and or
play already known medium. So now it has been made more of an optional
dependency that will only be imported when needed.

Also give a little more detail about what module failed to be imported
when a server fails to be instantiated