play: Introduce --raw option
So it isn't uncommon that some upstream server changes their api and
breaks all the worlds. And honestly, amt is not likely to ever be one of
the first to be updated. In that hopefully small window, allow the user
to use an arbitrary external player. Normally amt takes the
human-visible url, uses the servers api and translates it to a generic
video stream/url. Now with the `--raw` option, we'll pass the
human-visible url directly.

This functions similiarly to the user just calling their external player
directly. But this was we still maintain the ability to continue to the
next episode, progress tracking, tracker sync etc.