torrent: improve url for streaming torrents
While one could easily add a torrent file from a url, the ui for
streaming a specific file from the torrent was cumbersome. It required
the user to know the exact file id which is often a very long path name.

Now if there is only one file in the torrent, it will be auto selected
for streaming. Additionally, one can specify a number as a qet parameter
to the url to explicit a file by number. This probably won't work well
for special files but would ideally be much nicer in the common case